requirements.txt
Flask>=2.0.1
requests>=2.26.0
pytz>=2021.1
pyOpenSSL
certifi